Generate memorable names for the deployment namespace
Due to the fact that names can have at most 63 characters, including the namespace 
and instance name, together with a prefix has a high chance that the name is exceeding
this limit.
A random generated alphanumeric string is not memorable.
Restricting the max name length via admission validator is not user friendly either.