Set static name for deployment in deployment namespace
Using Helm's 'fullnameOverride' param we can set a static name.
The reason for doing so is that a lot of additional operators and controllers work with appending suffixes.
If we would continue using the same name as the instance CRD, we could encounter too long names and things may break.
Static names avoid this by setting a short-enough name.